refactor: Removed helpers/mocks/options
