+++ tests/helpers/mocks/index.ts
@@ -1,3 +1,2 @@
 export * from './components-double-colonized.js';
 export * from './entities.js';
-export * from './options.js';
