--- tests/helpers/mocks/options.ts
@@ -1,19 +1,0 @@
-import type { CodemodOptions, Options } from '../../../src/types/index.js';
-
-export const codemodOptions: CodemodOptions = {
-  componentStructure: 'flat',
-  convert: new Set(['components', 'routes', 'tests']),
-  folder: '',
-  projectRoot: 'tmp/my-monorepo',
-};
-
-export const options: Options = {
-  componentStructure: 'flat',
-  convert: {
-    components: true,
-    routes: true,
-    tests: true,
-  },
-  folder: '',
-  projectRoot: 'tmp/my-monorepo',
-};
